CHG migrate text editor to separate component
