+++ web/src/components/forums/RichTextEditor/Toolbar/EditorToolbar.jsx
