+++ web/src/components/forums/RichTextEditor/Toolbar/Embed/EmbedImage.jsx
